Split up ffi.cpp into several files
It was getting pretty big and would get worse over time, eg whenver
I get around to adding TLS support.
